# openapi.yaml
openapi: 3.1.0
info:
  title: IP Geolocation API
  description: |
    A RESTful API service for retrieving geolocation information based on IP addresses.
    
    This API provides detailed location data including country, region, city, coordinates,
    timezone, and ISP information for any valid IPv4 or IPv6 address.
    
    **Features:**
    - My IP lookup (automatically detects requester's IP address)
    - Single IP address lookup
    - Comprehensive error handling
    - Rate limiting support
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API service
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "healthy"
                timestamp: "2024-01-15T10:30:00Z"

  /ip:
    get:
      summary: Get geolocation for the requester's IP address
      description: |
        Retrieves geolocation information for the IP address of the client making the request.
        
        The API automatically detects the client's IP address by checking:
        - X-Forwarded-For header (for proxied/load-balanced requests)
        - X-Real-IP header
        - Direct client connection IP
        
        Supports both IPv4 and IPv6 addresses. Returns detailed location data
        including country, region, city, coordinates, timezone, and ISP information.
      operationId: getRequesterIpGeolocation
      tags:
        - Geolocation
      responses:
        '200':
          description: Successful response with geolocation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpGeolocationResponse'
              examples:
                success:
                  summary: Successful lookup
                  value:
                    ip: "203.0.113.1"
                    country: "United States"
                    country_code: "US"
                    region: "California"
                    region_code: "CA"
                    city: "San Francisco"
                    postal_code: "94102"
                    latitude: 37.7749
                    longitude: -122.4194
                    timezone: "America/Los_Angeles"
                    isp: "Example ISP"
                    org: "Example Organization"
                    asn: "AS12345"
                    query_timestamp: "2024-01-15T10:30:00Z"
        '404':
          description: IP address not found or no geolocation data available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "IP address not found"
                details:
                  - loc: ["path", "ip_address"]
                    msg: "No geolocation data available for IP address 192.168.1.1"
                    type: "ip_address_not_found_error"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Rate limit exceeded"
                details:
                  - loc: ["server"]
                    msg: "Too many requests. Please try again later."
                    type: "rate_limit_error"
        '500':
          description: Internal server error or external API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "An unexpected error occurred"
                details:
                  - loc: ["server"]
                    msg: "Internal server error"
                    type: "generic_error"
        '503':
          description: Service unavailable (external API timeout or unavailable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Service unavailable"
                details:
                  - loc: ["server"]
                    msg: "The geolocation service is temporarily unavailable. Please try again later."
                    type: "external_api_error"

  /ip/{ip_address}:
    get:
      summary: Get geolocation for a single IP address
      description: |
        Retrieves geolocation information for a single IP address.
        
        Supports both IPv4 and IPv6 addresses. Returns detailed location data
        including country, region, city, coordinates, timezone, and ISP information.
      operationId: getIpGeolocation
      tags:
        - Geolocation
      parameters:
        - name: ip_address
          in: path
          required: true
          description: The IP address to look up (IPv4 or IPv6)
          schema:
            type: string
            format: ipv4-or-ipv6
          examples:
            ipv4:
              value: "8.8.8.8"
              summary: IPv4 example
            ipv6:
              value: "2001:4860:4860::8888"
              summary: IPv6 example
      responses:
        '200':
          description: Successful response with geolocation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpGeolocationResponse'
              examples:
                success:
                  summary: Successful lookup
                  value:
                    ip: "8.8.8.8"
                    country: "United States"
                    country_code: "US"
                    region: "California"
                    region_code: "CA"
                    city: "Mountain View"
                    postal_code: "94043"
                    latitude: 37.4056
                    longitude: -122.0775
                    timezone: "America/Los_Angeles"
                    isp: "Google LLC"
                    org: "Google Public DNS"
                    asn: "AS15169"
                    query_timestamp: "2024-01-15T10:30:00Z"
        '422':
          description: Validation error (invalid IP address format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Validation error"
                details:
                  - loc: ["path", "ip_address"]
                    msg: "The provided IP address 'invalid-ip' is not a valid IPv4 or IPv6 address"
                    type: "validation_error"
        '404':
          description: IP address not found or no geolocation data available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "IP address not found"
                details:
                  - loc: ["path", "ip_address"]
                    msg: "No geolocation data available for IP address 192.168.1.1"
                    type: "ip_address_not_found_error"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Rate limit exceeded"
                details:
                  - loc: ["server"]
                    msg: "Too many requests. Please try again later."
                    type: "rate_limit_error"
        '500':
          description: Internal server error or external API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "An unexpected error occurred"
                details:
                  - loc: ["server"]
                    msg: "Internal server error"
                    type: "generic_error"
        '503':
          description: Service unavailable (external API timeout or unavailable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Service unavailable"
                details:
                  - loc: ["server"]
                    msg: "The geolocation service is temporarily unavailable. Please try again later."
                    type: "external_api_error"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Health status of the service
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Current server timestamp in ISO 8601 format
          example: "2024-01-15T10:30:00Z"

    IpGeolocationResponse:
      type: object
      required:
        - ip
        - country
        - country_code
        - query_timestamp
      properties:
        ip:
          type: string
          format: ipv4-or-ipv6
          description: The queried IP address
          example: "8.8.8.8"
        country:
          type: string
          description: Full country name
          example: "United States"
        country_code:
          type: string
          description: ISO 3166-1 alpha-2 country code
          minLength: 2
          maxLength: 2
          example: "US"
        region:
          type: string
          nullable: true
          description: Region or state name
          example: "California"
        region_code:
          type: string
          nullable: true
          description: Region or state code
          example: "CA"
        city:
          type: string
          nullable: true
          description: City name
          example: "Mountain View"
        postal_code:
          type: string
          nullable: true
          description: Postal or ZIP code
          example: "94043"
        latitude:
          type: number
          format: float
          nullable: true
          description: Latitude coordinate
          minimum: -90
          maximum: 90
          example: 37.4056
        longitude:
          type: number
          format: float
          nullable: true
          description: Longitude coordinate
          minimum: -180
          maximum: 180
          example: -122.0775
        timezone:
          type: string
          nullable: true
          description: IANA timezone identifier
          example: "America/Los_Angeles"
        isp:
          type: string
          nullable: true
          description: Internet Service Provider name
          example: "Google LLC"
        org:
          type: string
          nullable: true
          description: Organization name
          example: "Google Public DNS"
        asn:
          type: string
          nullable: true
          description: Autonomous System Number
          example: "AS15169"
        query_timestamp:
          type: string
          format: date-time
          description: Timestamp when the query was made (ISO 8601 format)
          example: "2024-01-15T10:30:00Z"

    ErrorDetail:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          description: Location of the error (field path)
          example: ["path", "ip_address"]
        msg:
          type: string
          description: Error message
          example: "Invalid IP address format"
        type:
          type: string
          description: Error type
          example: "validation_error"

    ErrorResponse:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
          description: Error message summary
          example: "Validation error"
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: List of error details
          example:
            - loc: ["path", "ip_address"]
              msg: "Invalid IP address format"
              type: "validation_error"

  securitySchemes: {}

tags:
  - name: Health
    description: Health check endpoints
  - name: Geolocation
    description: IP geolocation lookup endpoints